fixed styles, added heart animation

+++ view/chat/src/main.js
@@ -1,13 +1,13 @@
 import './styles.css';
 import { io } from 'socket.io-client';
 
-const socket = io('chat.liveshopping.gr/chat');
+const socket = io('localhost:3001/chat');
 
-// todo: move to another file
-const chatId = new URL(document.location.href).searchParams.get('id');
-
+const path = window.location.pathname;
+const chatId = path.substring(path.lastIndexOf('/') + 1);
+console.log('chatId: ' + chatId);
 if (!chatId) {
-  throw new Error('Chat room ID is not provided');
+  throw new Error('Chat room ID is not provided', path);
 }
 
 socket.on('connect', () => {
@@ -19,21 +19,47 @@ await socket.emitWithAck('create_room', {
   id: chatId,
 });
 
+let hotProducts = [];
+
 socket.on('receive_new_message', (data) => {
   const msg = JSON.parse(data);
-  renderNewMessage({
-    id: msg.id,
-    username: msg.user.name,
-    content: msg.content,
-    role: msg.user.role,
-    userId: msg.user.id,
-    reply: msg.reply?.content,
-    replyTo: msg.reply?.username,
-    currentRole: localStorage.getItem('role'),
-    likes: msg.likes,
-    type: msg.type,
-  });
-  scrollToBottom('chatMessagesContainer');
+  if (msg.type === 'text' && msg.content.startsWith('[hot-settings]')) {
+    var message = msg.content;
+    var ids = message.replace('[hot-settings]', '').split(',');
+    console.log(ids);
+
+    if (hotProducts !== ids) {
+      hotProducts = ids;
+
+      var hotProductItems = document.querySelectorAll(
+        "li[name='hot-product-item']",
+      );
+      hotProductItems.forEach(function (item) {
+        item.style.display = 'none';
+      });
+
+      ids.forEach(function (id) {
+        var hotItem = document.getElementById('hot-' + id);
+        if (hotItem) {
+          hotItem.style.display = 'block';
+        }
+      });
+    }
+  } else {
+    renderNewMessage({
+      id: msg.id,
+      username: msg.user.name,
+      content: msg.content,
+      role: msg.user.role,
+      userId: msg.user.id,
+      reply: msg.reply?.content,
+      replyTo: msg.reply?.username,
+      currentRole: localStorage.getItem('role'),
+      likes: msg.likes,
+      type: msg.type,
+    });
+    scrollToBottom('chatMessagesContainer');
+  }
 });
 
 function renderNewMessages(messages, toStart = false) {
@@ -76,17 +102,26 @@ function renderNewMessage(
     );
     return;
   }
+
+  if (data.content.startsWith('[hot-settings]')) {
+    return;
+  }
+
   const msgEl = document.createElement('div');
   msgEl.className = 'chat-msg-container';
 
   if (data.reply?.length && data.replyTo) {
     // todo: move to separate method
-    msgEl.innerHTML += `<div class="chat-msg__reply">
-                  <div class="reply-icon">
-                    <img src="assets/reply.svg" />
-                  </div>
+
+    msgEl.innerHTML += `<div class="chat-msg__reply-hold">
+                 
                   <div class="reply-content">
-                    <div class="reply-to-user">${data.replyTo}</div>
+                    <div class="reply-to-user">
+                    <div class="reply-icon">
+                    <img src="./assets/reply.svg" />
+                  </div>
+                 <div> ${data.replyTo}</div>
+                 </div>
                     <div class="reply-on-msg">
                       ${data.reply}
                     </div>
@@ -114,7 +149,7 @@ function renderNewMessage(
                   data.currentRole === 'vendor' ? '' : 'hide'
                 }">
                 <div class="block-msg-btn">
-                  <img src="assets/cancel.svg" />
+                  <img src="./assets/cancel.svg" />
                   </div>
                 </div>
                 `
@@ -124,9 +159,12 @@ function renderNewMessage(
                   data.likes.likedByMe ? 'active' : ''
                 }" data-liked-by-me="${data.likes.likedByMe}">
                     <i class="far fa-light fa-heart"></i>
-                    <div class="msg-like__counter" data-count="${
-                      data.likes.count
-                    }">${data.likes.count}</div>
+              
+                    
+                      <div class="msg-like__counter" data-count="${
+                        data.likes.count
+                      }">${data.likes.count !== 0 ? data.likes.count : ''}</div>
+                    
               </div>
               </div>`;
   const messagesEl = document.getElementById('chatMessages');
@@ -193,11 +231,37 @@ const onReady = async () => {
 
   if (userId) {
     document.querySelector('#selectRole').remove();
+    document.querySelector('#login-data').remove();
   }
 
   // load messages
   await fetchAndRenderMessages(chatId, false);
   scrollToBottom('chatMessagesContainer');
+
+  var localRole = localStorage.getItem('role');
+
+  var localSelectRole = document.getElementById('selectRole');
+
+  var localSelectedRole = localSelectRole ? localSelectRole.value : null;
+
+  if (localRole === 'vendor' || localSelectedRole === 'vendor') {
+    // hot products
+    updateHotProducts(false);
+
+    // setInterval(function() {
+    //   updateHotProducts(true);
+    // }, 1000);
+
+    var hotProductSelections = document.querySelectorAll(
+      "input[name='hot-product-selection']",
+    );
+    hotProductSelections.forEach(function (hotProductSelection) {
+      hotProductSelection.addEventListener('change', function () {
+        console.log(this);
+        updateHotProducts(true);
+      });
+    });
+  }
 };
 
 if (document.readyState !== 'loading') {
@@ -208,8 +272,10 @@ if (document.readyState !== 'loading') {
 
 // handle form submit
 const form = document.getElementById('chatForm');
+const loginWrapper = document.getElementById('login-data');
 form.addEventListener('submit', async (e) => {
   e.preventDefault();
+  loginWrapper.style.display = 'none';
   const formData = new FormData(e.target);
   const content = formData.get('msg');
   let role;
@@ -226,6 +292,7 @@ form.addEventListener('submit', async (e) => {
 });
 
 async function sendMessage({ type, role, content } = {}) {
+  console.log({ type, role, content });
   if (!content) {
     console.error('error on msg submit');
     return;
@@ -237,6 +304,7 @@ async function sendMessage({ type, role, content } = {}) {
     role: role || 'user',
     type,
   };
+
   const curReply = document.getElementById('curReply');
   const msgInput = form.querySelector('input[name="msg"]');
   const userId = localStorage.getItem('chatUserId');
@@ -247,6 +315,8 @@ async function sendMessage({ type, role, content } = {}) {
     payload.userId = userId;
   } else if (username) {
     payload.username = username;
+  } else if (role === 'vendor') {
+    payload.username = 'Vendor';
   } else {
     // todo: handle error
     console.error('error on msg submit');
@@ -256,27 +326,33 @@ async function sendMessage({ type, role, content } = {}) {
   if (curReply && curReply.dataset.replyToMsgId) {
     payload.replyTo = +curReply.dataset.replyToMsgId;
   }
+  console.log(payload);
 
   const response = await socket.emitWithAck('send_message', payload);
   const parsedRes = JSON.parse(response);
 
+  console.log(response);
+
   if (parsedRes.success) {
-    localStorage.setItem('chatUserId', parsedRes.data.userId);
-    userInput?.remove();
-    msgInput.value = '';
-    deleteCurrentReply();
-    scrollToBottom('chatMessagesContainer');
+    if (content.startsWith('[hot-settings]')) {
+    } else {
+      localStorage.setItem('chatUserId', parsedRes.data.userId);
+      userInput?.remove();
+      msgInput.value = '';
+      deleteCurrentReply();
+      scrollToBottom('chatMessagesContainer');
 
-    if (role) {
-      // todo: should be replaced for prod
-      payload.role = role;
-      localStorage.setItem('role', role + '');
-      document.querySelector('#selectRole')?.remove();
+      if (role) {
+        // todo: should be replaced for prod
+        payload.role = role;
+        localStorage.setItem('role', role + '');
+        document.querySelector('#selectRole')?.remove();
 
-      if (role === 'vendor') {
-        document.querySelectorAll('.block-msg').forEach((msg) => {
-          msg.classList.remove('hide');
-        });
+        if (role === 'vendor') {
+          document.querySelectorAll('.block-msg').forEach((msg) => {
+            msg.classList.remove('hide');
+          });
+        }
       }
     }
   }
@@ -309,7 +385,7 @@ document.getElementById('chatMessages').addEventListener('click', (e) => {
 
     const counterEl = likeContainer.querySelector('.msg-like__counter');
     const count = +counterEl.dataset.count + (likedByMe ? -1 : 1);
-    counterEl.innerHTML = count;
+    counterEl.innerHTML = count !== 0 ? count : '';
     counterEl.dataset.count = count;
 
     socket.emit('like_message', {
@@ -356,23 +432,30 @@ function holdReply(e) {
     reply.classList.remove('hide');
     reply.dataset.replyToMsgId = closestMsg.dataset.id;
     reply.innerHTML = `
-              <div class="reply-icon">
-                    <img src="assets/reply.svg" />
-                  </div>
+              
                   <div class="reply-content">
-                    <div class="reply-to-user">${msgUserEl.innerText}</div>
+                    <div class="reply-to-user">
+                    <div class="reply-icon">
+                    <img src="./assets/reply.svg" />
+                  </div>
+                  <div>${msgUserEl.innerText}</div>
+                  </div>
                     <div class="reply-on-msg">
-                      ${msgContent}
+                      ${
+                        msgContent.length < 30
+                          ? msgContent
+                          : msgContent.substring(0, 27) + '...'
+                      }
                     </div>
                   </div>
                 </div>
               <div class="reply-cancel" id="cancelReply">
-                <img src="assets/cancel.svg" />
+                <img src="./assets/cancel.svg" />
               </div>
               </div>`;
 
     // handle add && cancel reply
-    reply.querySelector('#cancelReply').addEventListener('click', (e) => {
+    reply.querySelector('#cancelReply').addEventListener('click', () => {
       deleteCurrentReply();
     });
   }
@@ -439,13 +522,65 @@ function scrollToBottom(id) {
 
 function animateHearts() {
   const heartsParent = document.querySelector('.chat__send_like');
-  heartsParent.classList.add('animate');
-  setTimeout(() => {
-    heartsParent.classList.remove('animate');
-  }, 1000);
+
+  const heart = document.createElement('i');
+  heart.className = 'far fa-light video-hearts fa-heart';
+  heart.style.top = 0;
+  heart.style.left = '-2vw';
+  heartsParent.appendChild(heart);
+  const animationScenarios = [
+    [
+      { opacity: 0, left: '-2vw', top: '0vh' },
+      { opacity: 1, left: '2vw', top: '-5vh' },
+      { opacity: 1, left: '-2vw', top: '-15vh' },
+      { opacity: 1, left: '2vw', top: '-25vh' },
+      { opacity: 1, left: '-2vw', top: '-35vh' },
+      { opacity: 1, left: '2vw', top: '-45vh' },
+      { opacity: 0, left: '-2vw', top: '-55vh' },
+    ],
+    [
+      { opacity: 0, left: '-3vw', top: '0vh' },
+      { opacity: 1, left: '1vw', top: '-8vh' },
+      { opacity: 1, left: '-3vw', top: '-18vh' },
+      { opacity: 1, left: '1vw', top: '-28vh' },
+      { opacity: 1, left: '-2vw', top: '-38vh' },
+      { opacity: 1, left: '1vw', top: '-48vh' },
+      { opacity: 0, left: '-1vw', top: '-58vh' },
+    ],
+    [
+      { opacity: 0, left: '-1vw', top: '0vh' },
+      { opacity: 1, left: '1vw', top: '-3vh' },
+      { opacity: 1, left: '-1vw', top: '-13vh' },
+      { opacity: 1, left: '1vw', top: '-23vh' },
+      { opacity: 1, left: '-1vw', top: '-33vh' },
+      { opacity: 1, left: '1vw', top: '-43vh' },
+      { opacity: 0, left: '-1vw', top: '-53vh' },
+    ],
+  ];
+
+  function getRandomScenario() {
+    const randomIndex = Math.floor(Math.random() * animationScenarios.length);
+    return animationScenarios[randomIndex];
+  }
+  const randomScenario = getRandomScenario();
+  const animation = heart.animate(randomScenario, {
+    duration: 2700,
+    fill: 'forwards',
+  });
+  animation.addEventListener('finish', () => {
+    heartsParent.removeChild(heart);
+  });
+  // animation.onfinish(() => {
+  //   document.body.removeChild(heart);
+  // });
+  // heartsParent.classList.add('animate');
+  // setTimeout(() => {
+  //   heartsParent.classList.remove('animate');
+  // }, 1000);
 }
 
-document.querySelector('#sendLikeBtn')?.addEventListener('click', () => {
+document.querySelector('#sendLikeBtn')?.addEventListener('click', (e) => {
+  e.preventDefault();
   const userId = localStorage.getItem('chatUserId');
 
   socket.emit('send_like', {
@@ -455,8 +590,8 @@ document.querySelector('#sendLikeBtn')?.addEventListener('click', () => {
   animateHearts();
 });
 
-socket.on('send_like', (data) => {
-  const parsed = JSON.parse(data);
+socket.on('send_like', () => {
+  // const parsed = JSON.parse(data);
   animateHearts();
 });
 
@@ -489,3 +624,38 @@ document.addEventListener('click', (e) => {
     document.querySelector('#stickersContainer').classList.add('hide');
   }
 });
+
+function updateHotProducts(isSendMessage) {
+  var productIds = document.querySelectorAll(
+    "input[name='hot-product-selection']:checked",
+  );
+  var ids = [];
+  productIds.forEach(function (productId) {
+    var sThisVal = productId.checked ? productId.value : '';
+    ids.push(sThisVal);
+  });
+  console.log(ids);
+
+  var hotProductItems = document.querySelectorAll(
+    "li[name='hot-product-item']",
+  );
+  hotProductItems.forEach(function (item) {
+    item.style.display = 'none';
+  });
+
+  ids.forEach(function (id) {
+    var hotItem = document.getElementById('hot-' + id);
+    if (hotItem) {
+      hotItem.style.display = 'block';
+    }
+  });
+
+  if (isSendMessage) {
+    // send message to all clients
+    sendMessage({
+      type: 'text',
+      content: '[hot-settings]' + ids,
+      role: 'vendor',
+    }).catch((e) => console.log(e));
+  }
+}
